requirements.txt
numpy
pandas
matplotlib
opencv-python
Pillow
requests
Flask
discord.py
imageio